backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
google-generativeai
cryptography
Pillow
python-multipart